pyproject.toml
[project]
name = "pyrlottie"
version = "2024.0.1"
description = "Convert .tgs and .lottie to .webp or .gif using rlottie."
authors = [
	{name = "FredHappyface", email = "FredHappyface@users.noreply.github.com"}
]
classifiers = [
	"License :: OSI Approved :: MIT License",
	"Development Status :: 5 - Production/Stable",
	"Intended Audience :: Developers",
	"Intended Audience :: Education",
	"Natural Language :: English",
	"Operating System :: OS Independent",
	"Programming Language :: Python :: Implementation :: CPython",
	"Topic :: Software Development :: Libraries :: Python Modules",
	"Topic :: Utilities",
	"Topic :: Multimedia :: Graphics",
]
homepage = "https://github.com/FHPythonUtils/PyRlottie"
repository = "https://github.com/FHPythonUtils/PyRlottie"
documentation = "https://github.com/FHPythonUtils/PyRlottie/blob/master/README.md"
readme = "README.md"
dependencies = [
	"attrs>=23.2.0",
	"Pillow",
	"numpy",
	"deprecation>=2.1.0",
]

[project.optional-dependencies]
dev = [
	"imgcompare>=2.0.1",
	"pytest>=8.1.1",
	"handsdown>=2.1.0",
	"coverage>=7.4.4",
	"ruff>=0.3.3",
	"pyright>=1.1.354",
]

[build-system]
requires = [
	"py-build-cmake~=0.2.0a12",
	"conan>=2.0",
	"patch>=1.16",
]
build-backend = "py_build_cmake.build"

[tool.py-build-cmake.module]
name = "pyrlottie"
directory = "src"

[tool.py-build-cmake.sdist]
include = [
	"scripts/*",
	"patches/*",
	"CMakeLists.txt",
	"rlottie/*",
	"libwebp/*",
	"conanfile.py",
]

[tool.py-build-cmake.cmake]
build_type = "Release"
source_path = "."
build_args = ["-j"]
install_components = ["python_module"]
python_abi = 'abi3'
abi3_minimum_cpython_version = 38

[tool.ruff]
line-length = 100
indent-width = 4
target-version = "py38"

[tool.ruff.lint]
select = ["ALL"]
ignore = [
	"ANN101",  # type annotation for self in method
	"COM812",  # enforce trailing comma
	"D2",  # pydocstyle formatting
	"ISC001",
	"N",  # pep8 naming
	"PLR09",  # pylint refactor too many
	"TCH",  # type check blocks
	"W191"  # ignore this to allow tabs
]
fixable = ["ALL"]

[tool.ruff.lint.per-file-ignores]
"**/{tests,docs,tools}/*" = ["D", "S101", "E402"]

[tool.ruff.lint.flake8-tidy-imports]
ban-relative-imports = "all" # Disallow all relative imports.

[tool.ruff.format]
indent-style = "tab"
docstring-code-format = true
line-ending = "lf"

[tool.pyright]
venvPath = "."
venv = ".venv"

[tool.coverage.run]
branch = true

[tool.tox]
legacy_tox_ini = """
[tox]
env_list =
	py312
	py311
	py310
	py39
	py38

[testenv]
deps =
	imgcompare
	pytest
commands = pytest tests
"""
